refactor: move protocol helpers to utils
