+++ src/massconfigmerger/result_processor.py
@@ -14,6 +14,7 @@ from urllib.parse import urlparse, parse_qs, urlencode, urlunparse, parse_qsl
 
 from .config import Settings, load_config
 from .tester import NodeTester
+from .utils import is_valid_config, parse_configs_from_text
 
 DEFAULT_CONFIG_FILE = Path(__file__).resolve().with_name("config.yaml")
 try:
