+++ tests/test_parse_configs.py
@@ -1,11 +1,11 @@
 import base64
 
-from massconfigmerger import aggregator_tool
+from massconfigmerger import utils
 
 
 def test_multiple_links_same_line():
     text = "vmess://a vless://b"
-    result = aggregator_tool.parse_configs_from_text(text)
+    result = utils.parse_configs_from_text(text)
     assert "vmess://a" in result
     assert "vless://b" in result
     assert len(result) == 2
@@ -31,21 +31,21 @@ def test_extract_all_protocols():
             "wireguard://o",
         ]
     )
-    result = aggregator_tool.parse_configs_from_text(text)
+    result = utils.parse_configs_from_text(text)
     assert len(result) == 15
     for item in text.split():
         assert item in result
 
 
 def test_parse_oversized_line(caplog):
-    big_line = "A" * (aggregator_tool.MAX_DECODE_SIZE + 1)
+    big_line = "A" * (utils.MAX_DECODE_SIZE + 1)
     with caplog.at_level('DEBUG'):
-        result = aggregator_tool.parse_configs_from_text(big_line)
+        result = utils.parse_configs_from_text(big_line)
     assert result == set()
     assert "Skipping oversized base64 line" in caplog.text
 
 
 def test_urlsafe_base64_line():
     encoded = base64.urlsafe_b64encode(b"vmess://a  >").decode()
-    result = aggregator_tool.parse_configs_from_text(encoded)
+    result = utils.parse_configs_from_text(encoded)
     assert result == {"vmess://a"}
